# [WIP] basecase for flint_mpn_mul and flint_mpn_mul_n