--- a/CMake/cmake_config.h.in
+++ b/CMake/cmake_config.h.in
@@ -20,6 +20,8 @@
 
 #cmakedefine FLINT_HAVE_FFT_SMALL
 
+#cmakedefine01 FLINT_KNOW_STRONG_ORDER
+
 #ifdef _MSC_VER
 # if defined(FLINT_BUILD_DLL)
 #  define FLINT_DLL __declspec(dllexport)