--- a/Makefile.in
+++ b/Makefile.in
@@ -276,7 +276,7 @@ $(1)_TEST_SOURCES := $(wildcard $(SRC_DIR)/$(1)/test/main.c)
 endef
 _TEST_SOURCES := $(wildcard $(SRC_DIR)/test/main.c)
 $(foreach dir, $(DIRS), $(eval $(call xxx_TEST_SOURCES,$(dir))))
-TEST_SOURCES := $(foreach dir,$(DIRS),$($(dir)_TEST_SOURCES)) $(_TEST_SOURCES) $(fq_zech_vec_TEST_SOURCES)
+TEST_SOURCES := $(_TEST_SOURCES) $(foreach dir,$(DIRS),$($(dir)_TEST_SOURCES)) $(fq_zech_vec_TEST_SOURCES)
 # NOTE: We do not add CPP files to C files in SOURCES in order to not screw up
 # the `patsubst' in the object files.
 ifneq ($(WANT_NTL), 0)
@@ -334,7 +334,7 @@ $(foreach dir, $(DIRS), $(eval $(call xxx_TESTS,$(dir))))
 ifneq ($(WANT_NTL), 0)
 interfaces_TESTS := $(BUILD_DIR)/interfaces/test/t-NTL-interface$(EXEEXT)
 endif
-TESTS := $(foreach dir,$(DIRS),$($(dir)_TESTS)) $(_TESTS) $(interfaces_TESTS)
+TESTS := $(_TESTS) $(foreach dir,$(DIRS),$($(dir)_TESTS)) $(interfaces_TESTS)
 
 define xxx_TUNES
 $(1)_TUNES := $(patsubst $(SRC_DIR)/%.c,$(BUILD_DIR)/%$(EXEEXT),$($(1)_TUNE_SOURCES))