--- a/src/acb/approx_dot.c
+++ b/src/acb/approx_dot.c
@@ -605,14 +605,14 @@ acb_approx_dot(acb_t res, const acb_t initial, int subtract, acb_srcptr x, slong
                             x1 = ARF_NOPTR_D(xm)[1];
                             y0 = ARF_NOPTR_D(ym)[0];
                             y1 = ARF_NOPTR_D(ym)[1];
-                            flint_mpn_mul_2x2(u3, u2, u1, u0, x1, x0, y1, y0);
+                            FLINT_MPN_MUL_2X2(u3, u2, u1, u0, x1, x0, y1, y0);
                         }
                         else if (xn == 1)
                         {
                             x0 = ARF_NOPTR_D(xm)[0];
                             y0 = ARF_NOPTR_D(ym)[0];
                             y1 = ARF_NOPTR_D(ym)[1];
-                            flint_mpn_mul_2x1(u3, u2, u1, y1, y0, x0);
+                            FLINT_MPN_MUL_2X1(u3, u2, u1, y1, y0, x0);
                             u0 = 0;
                         }
                         else
@@ -620,7 +620,7 @@ acb_approx_dot(acb_t res, const acb_t initial, int subtract, acb_srcptr x, slong
                             x0 = ARF_NOPTR_D(xm)[0];
                             x1 = ARF_NOPTR_D(xm)[1];
                             y0 = ARF_NOPTR_D(ym)[0];
-                            flint_mpn_mul_2x1(u3, u2, u1, x1, x0, y0);
+                            FLINT_MPN_MUL_2X1(u3, u2, u1, x1, x0, y0);
                             u0 = 0;
                         }