--- a/src/mpn_extras/profile/p-mul_n.c
+++ b/src/mpn_extras/profile/p-mul_n.c
@@ -13,7 +13,7 @@
 #include "mpn_extras.h"
 #include "profiler.h"
 
-#define MAXN 20
+#define MAXN 15
 
 int main()
 {