It appears that we can do better than GMP's basecase multiplication using a codelet approach (so there is no loop overhead) and by doing columnwise updates (i.e. dot products) instead of rowwise updates (i.e. addmul_1's).

On my laptop (Zen3), I get timings like this:

```
     mpn_mul_n   flint_mpn_mul_n  speedup
n = 1    3.98e-09    2.13e-09   1.869x
n = 2    4.92e-09    2.38e-09   2.067x
n = 3    7.93e-09    4.04e-09   1.963x
n = 4    1.21e-08    7.32e-09   1.653x
n = 5    1.62e-08    1.17e-08   1.385x
n = 6    2.37e-08    1.77e-08   1.339x
n = 7    2.92e-08    2.52e-08   1.159x
n = 8    3.68e-08    3.02e-08   1.219x
n = 9    4.57e-08    3.78e-08   1.209x
n = 10    5.82e-08    4.72e-08   1.233x
n = 11    6.65e-08    5.73e-08   1.161x
n = 12    7.82e-08    6.64e-08   1.178x
n = 13    9.05e-08    9.16e-08   0.988x
n = 14    1.11e-07    1.1e-07   1.009x
n = 15    1.21e-07    1.22e-07   0.992x
```

This is a WIP because we need to confirm that this isn't massively slower on other architectures, for which we should put in some logic to fall back to ``mpn_mul_n``.

The speedup seems to persist up to n = 20 or so but I've only done up to n = 12 here since i-cache misses might become an issue. I'm hoping that slightly bigger sizes and unbalanced multiplications can be done as efficiently using a block strategy (e.g. 16x16 using four 8x8 products). Codelets + Karatsuba should do well eventually too.

Squaring and unbalanced multiplications to be done later. The macros and inline fixed-size multiplications should also later be exported to ``mpn_extras.h`` since they would be useful in various places (fully inlining a fixed-sized multiplication in the calling code will give further speedups).