@albinahlback 2023-12-14
Looks really nice!

I suppose you used GCC to benchmark? In that case we should really want to optimize `longlong.h` as (IIRC) GMP bases its `mpn_mul_basecase` off of assembly for a lot of processors, and this code should only perform better if `add_ssaaa` and friends are somewhat optimal.