@fredrik-johansson 2023-12-14
OK, so here is what's going on (thanks to Dan for some feedback).

GMP's inner loops do 2 adds for each mul in the addmul_1 loop. This is possible thanks to the mulx/adox/adcx instructions which allow maintaining two separate carry chains.

The dot product method does 3 adds for each mul which is asymptotically slower, so basically all the speedup comes from avoiding loop overhead.

[This project](https://github.com/krisk0/broadwell_multiplication) has some fixed-size asm routines based on the mulx/adox/adcx method. Timing comparison for 8x8:

    mpn_mul_n    flint_mpn_mul_8       mul8_zen
    3.64e-08        2.96e-08           2.71e-08

So that would give us another 10% speedup. I don't see how to produce that kind of code from C except by using huge blocks of inline assembly though.

I'm thinking we should really bite the bullet and write our own assembly code generator for basic mpn routines. We can't use the one I just linked to, because it's LGPL3, unfortunately. 

Anyway, the code in this PR could perhaps be a decent intermediate solution.

We could perhaps do a bit better by playing Tetris with the partial products and doing fewer, wider multi-limb additions, but that's going to get ugly...