@fredrik-johansson 2023-12-15
Now using basecases for both `flint_mpn_mul` and `flint_mpn_mul_n`.

Current results for `flint_mpn_mul_n`:

```
n = 1    3.92e-09    2.34e-09   1.675x
n = 2    4.63e-09    2.35e-09   1.970x
n = 3    7.85e-09    3.87e-09   2.028x
n = 4    1.19e-08    7.19e-09   1.655x
n = 5    1.58e-08    1.13e-08   1.398x
n = 6    2.35e-08    1.67e-08   1.407x
n = 7    2.87e-08    2.27e-08   1.264x
n = 8    3.61e-08    2.98e-08   1.211x
n = 9    4.48e-08    3.73e-08   1.201x
n = 10    5.72e-08    4.62e-08   1.238x
n = 11    6.55e-08    5.58e-08   1.174x
n = 12    7.69e-08    6.5e-08   1.183x
n = 13    8.9e-08    8.99e-08   0.990x
n = 14    1.07e-07    1.08e-07   0.991x
n = 15    1.19e-07    1.2e-07   0.992x
n = 16    1.34e-07    1.34e-07   1.000x
n = 17    1.5e-07    1.5e-07   1.000x
n = 18    1.74e-07    1.75e-07   0.994x
n = 19    1.83e-07    1.9e-07   0.963x
n = 20    1.95e-07    2.01e-07   0.970x
```

Current results for `flint_mpn_mul` (speedups vs `mpn_mul`):

```
n = 1    1.148x  
n = 2    1.642x  1.764x  
n = 3    1.436x  1.694x  1.638x  
n = 4    1.631x  1.762x  1.596x  1.493x  
n = 5    1.348x  1.623x  1.434x  1.359x  1.310x  
n = 6    1.583x  1.594x  1.464x  1.405x  1.314x  1.311x  
n = 7    1.621x  1.465x  1.387x  1.262x  1.244x  1.209x  1.218x  
n = 8    1.608x  1.499x  1.333x  1.235x  1.230x  1.186x  1.167x  1.168x  
n = 9    1.612x  1.388x  1.272x  1.210x  1.167x  1.172x  1.153x  1.169x  1.188x  
n = 10    1.682x  1.340x  1.291x  1.218x  1.235x  1.212x  1.211x  1.228x  1.237x  1.205x  
n = 11    1.194x  0.925x  0.954x  0.960x  0.967x  0.969x  0.973x  0.978x  0.977x  0.984x  1.150x  
n = 12    1.157x  0.931x  0.949x  0.959x  0.969x  0.974x  0.977x  0.980x  0.980x  0.982x  0.983x  1.172x  
n = 13    1.159x  0.933x  0.955x  0.961x  0.963x  0.973x  0.978x  0.986x  0.984x  0.984x  0.979x  0.992x  0.977x  
n = 14    1.155x  0.942x  0.963x  0.965x  0.974x  0.976x  0.977x  0.984x  0.985x  0.990x  0.988x  0.987x  0.990x  0.982x  
n = 15    1.155x  0.944x  0.956x  0.969x  0.974x  0.978x  0.983x  0.983x  0.984x  0.985x  0.987x  0.987x  0.981x  0.991x  0.983x  
n = 16    1.124x  0.944x  0.966x  0.968x  0.976x  0.978x  0.984x  0.984x  0.987x  0.987x  0.986x  0.990x  0.991x  1.000x  0.984x  0.993x  
```

I've put in code for all unbalanced multiplications up to n = 10. We could clearly go even farther, especially for nx1 and nx2 products. Though first I'd like to see some results on macrobenchmarks, and on other machines...

The dispatch is now done by looking up a function in a table of function pointers. We could even consider defining ``flint_mpn_mul`` and ``flint_mpn_mul_n`` as inline functions so that there is even less overhead jumping into the right subroutine.