@fredrik-johansson 2023-12-16
It's difficult to find examples where this really makes a difference.

I believe this is because we already try hard to avoid doing few-word integer operations, early on switching to KS multiplication for polynomials and so on, and where we do lots of few-limb multiplications, e.g. for arb, there is a lot of overhead around each call to ``mpn_mul`` anyway. A couple of Arb examples which do see a tiny speedup:

```build/examples/hilbert_matrix 120```

Old: 0.51 seconds
New: 0.47 seconds

```build/examples/poly_roots w 100```

Old: 1.42 seconds
New: 1.29 seconds

Of course, my intention is to rewrite much of the basic arb arithmetic to use truncating multiplications and to remove as much as the remaining overhead as possible, which should give bigger improvements...